Panta rhei.
git-svn-id: https://svn.micromata.de/svn/projectforge/webapp/trunk@12393 be79319d-3443-0410-9b0a-b0c1b946f33b

+++ src/test/java/org/projectforge/ldap/LdapUserDaoTest.java
@@ -46,21 +46,25 @@ public class LdapUserDaoTest extends LdapRealTestBase
     ldapUserDao = new LdapUserDao();
     ldapOrganizationalUnitDao = new LdapOrganizationalUnitDao();
     ldapUserDao.ldapConnector = ldapOrganizationalUnitDao.ldapConnector = ldapConnector;
-    ldapOrganizationalUnitDao.createIfNotExist(PATH, "Test area for tests of ProjectForge.");
-    ldapOrganizationalUnitDao.createIfNotExist("deactivated", "for deactivated objects.", PATH);
+    if (ldapConfig != null) {
+      ldapOrganizationalUnitDao.createIfNotExist(PATH, "Test area for tests of ProjectForge.");
+      ldapOrganizationalUnitDao.createIfNotExist("deactivated", "for deactivated objects.", PATH);
+    }
   }
 
   @After
   public void tearDown()
   {
-    ldapOrganizationalUnitDao.deleteIfExists("deactivated", PATH);
-    ldapOrganizationalUnitDao.deleteIfExists(PATH);
+    if (ldapConfig != null) {
+      ldapOrganizationalUnitDao.deleteIfExists("deactivated", PATH);
+      ldapOrganizationalUnitDao.deleteIfExists(PATH);
+    }
   }
 
   @Test
   public void createAuthenticateAndDeleteUser()
   {
-    if (cfg == null) {
+    if (ldapConfig == null) {
       log.info("No LDAP server configured for tests. Skipping test.");
       return;
     }
@@ -88,7 +92,7 @@ public class LdapUserDaoTest extends LdapRealTestBase
   @Test
   public void activateAndDeactivateUser()
   {
-    if (cfg == null) {
+    if (ldapConfig == null) {
       log.info("No LDAP server configured for tests. Skipping test.");
       return;
     }
